requirements.txt
dify_plugin>=0.2.1,<0.3.0
nacos-maintainer-sdk-python==0.1.1
mcp==1.9.2
